+++ lib/src/processor/processor/ephemeral_message_processor.dart
@@ -1,9 +1,9 @@
 import 'package:mark/src/message/log_message.dart';
-import 'package:mark/src/processor/message_processor.dart';
-import 'package:mark/src/processor/print_message_printer.dart'
-    if (dart.library.html) 'package:mark/src/processor/web_message_printer.dart'
-    if (dart.library.io) 'package:mark/src/processor/io_message_printer.dart';
-import 'package:mark/src/processor/string_message_formatter_mixin.dart';
+import 'package:mark/src/processor/formatter/string_message_formatter_mixin.dart';
+import 'package:mark/src/processor/processor/message_processor.dart';
+import 'package:mark/src/processor/processor/print_message_printer.dart'
+    if (dart.library.html) 'package:mark/src/processor/processor/web_message_printer.dart'
+    if (dart.library.io) 'package:mark/src/processor/processor/io_message_printer.dart';
 
 /// A [MessageProcessor] that prints the formatted message to the console, using
 /// platform-specific methods.