+++ lib/src/processor/processor/io_message_printer.dart
@@ -1,7 +1,8 @@
 import 'dart:io';
 
 import 'package:mark/src/message/log_message.dart';
-import 'package:mark/src/processor/print_message_printer.dart' as printer;
+import 'package:mark/src/processor/processor/print_message_printer.dart'
+    as printer;
 
 void printMessage(LogMessage message, String data) {
   stdout.hasTerminal