+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] – 2023-02-04
+### Added
+- `IdentityMessageFormatterMixin` that acts as an identity function for messages
+
+### Changed
+- [BREAKING] `LogMessage.matchPrimitive`'s second argument is now a nullary function instead of being unary
+
+
 ## [0.1.0] – 2023-01-08
 ### Added
 - Initial version